refactor transaction review summary
