+++ app/components/TransactionReview/TransactionReviewSummary/index.test.js
@@ -1,17 +1,21 @@
 import React from 'react';
-import TransactionReview from './';
+import TransactionReviewSummary from './';
 import configureMockStore from 'redux-mock-store';
 import { shallow } from 'enzyme';
 
 const mockStore = configureMockStore();
 
-describe('TransactionReview', () => {
+describe('TransactionReviewSummary', () => {
 	it('should render correctly', () => {
 		const initialState = {
 			engine: {
 				backgroundState: {
-					PreferencesController: {
-						selectedAddress: '0x2'
+					TokenRatesController: {
+						contractExchangeRates: {}
+					},
+					CurrencyRateController: {
+						currentCurrency: 'usd',
+						conversionRate: 0.1
 					}
 				}
 			},
@@ -21,7 +25,7 @@ describe('TransactionReview', () => {
 		};
 
 		const wrapper = shallow(
-			<TransactionReview
+			<TransactionReviewSummary
 				navigation={{ state: { params: {} } }}
 				transactionData={{ amount: 0, gas: 0, gasPrice: 1, from: '0x0' }}
 			/>,
